Use an interface for kraken, it will help for testing and grpc kraken
