+++ cmd/schedules/main.go
@@ -17,6 +17,7 @@ import (
 	"github.com/CanalTP/gormungandr"
 	"github.com/CanalTP/gormungandr/auth"
 	"github.com/CanalTP/gormungandr/internal/schedules"
+	"github.com/CanalTP/gormungandr/kraken"
 	cache "github.com/patrickmn/go-cache"
 	"github.com/rafaeljesus/rabbus"
 
@@ -111,7 +112,7 @@ func main() {
 	})
 	logger.Info("starting schedules")
 
-	kraken := gormungandr.NewKraken("default", config.Kraken, config.Timeout)
+	kraken := kraken.NewKrakenZMQ("default", config.Kraken, config.Timeout)
 	authOption := schedules.SkipAuth()
 	var statPublisher *auth.StatPublisher
