+++ prometheus.go
@@ -35,31 +35,6 @@ var (
 	},
 	)
 
-	krakenDurations = prometheus.NewHistogramVec(prometheus.HistogramOpts{
-		Namespace: "gormungandr",
-		Subsystem: "kraken",
-		Name:      "durations_seconds",
-		Help:      "kraken request latency distributions.",
-		Buckets:   prometheus.ExponentialBuckets(0.001, 1.5, 25),
-	},
-		[]string{"api"},
-	)
-
-	krakenErrors = prometheus.NewCounterVec(prometheus.CounterOpts{
-		Namespace: "gormungandr",
-		Subsystem: "kraken",
-		Name:      "errors_count",
-		Help:      "kraken request errors count",
-	},
-		[]string{"api"},
-	)
-	krakenInFlight = prometheus.NewGauge(prometheus.GaugeOpts{
-		Namespace: "gormungandr",
-		Subsystem: "kraken",
-		Name:      "in_flight",
-		Help:      "current number of request being called",
-	},
-	)
 	sqlDurations = prometheus.NewHistogram(prometheus.HistogramOpts{
 		Namespace: "gormungandr",
 		Subsystem: "sql",
@@ -72,9 +47,6 @@ var (
 
 func init() {
 	prometheus.MustRegister(httpDurations)
-	prometheus.MustRegister(krakenDurations)
-	prometheus.MustRegister(krakenErrors)
-	prometheus.MustRegister(krakenInFlight)
 	prometheus.MustRegister(httpInFlight)
 	prometheus.MustRegister(sqlDurations)
 	sql.Register("postgresInstrumented", sqlhooks.Wrap(&pq.Driver{}, &instrumentHook{}))
@@ -88,33 +60,6 @@ func InstrumentGin() gin.HandlerFunc {
 		httpInFlight.Dec()
 		observer := httpDurations.With(prometheus.Labels{"handler": c.HandlerName(), "code": strconv.Itoa(c.Writer.Status())})
 		observer.Observe(time.Since(begin).Seconds())
-	}
-}
-
-type RequestObserver struct {
-	api   string
-	begin time.Time
-}
-
-func (o RequestObserver) Finish() {
-	krakenDurations.With(prometheus.Labels{"api": o.api}).
-		Observe(time.Since(o.begin).Seconds())
-	krakenInFlight.Dec()
-}
-
-func (o KrakenObserver) OnError(api string, err error) {
-	krakenErrors.With(prometheus.Labels{"api": api}).
-		Inc()
-}
-
-type KrakenObserver struct {
-}
-
-func (o KrakenObserver) StartRequest(kraken *Kraken, api string) RequestObserver {
-	krakenInFlight.Inc()
-	return RequestObserver{
-		api:   api,
-		begin: time.Now(),
 	}
 }
